Added first prototype of packet sequence numbering.
Currently using ByteBuffer methods to convert the packet number into a byte array but not sure if we are allowed to use it for the final implementation or whether we can use manual bit shifting instead.

The packet number is being added to the payload because we cannot access the header using the provided jar files (as far as I know) Because the packet number (of size 8 bytes) is being added to the payload (which is already 512 bytes) we end up with a total payload size of 520 bytes.